Changed price data type to numeric instead of float. Limit price input to two decimal point

+++ app.py
@@ -12,7 +12,7 @@ class Expense(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(50), nullable=False)
     category = db.Column(db.String(50), nullable=False)
-    price = db.Column(db.Float, nullable=False)
+    price = db.Column(db.Numeric(10,2), nullable=False)
 
 @app.route('/')
 def index():
